localize functions and reduce prop drilling
