+++ src/screens/Project/TaskManagement/TicketDialog/Comments/CommentContent.tsx
@@ -1,10 +1,27 @@
+import { updateComment } from 'utils/api/tickets'
+
 export const CommentContent = ({
+  comment,
   editMode,
   editedComment,
   setEditedComment,
-  saveUpdatedComment,
-  content,
+  toggleFetchComments,
+  toggleEditMode,
 }) => {
+  const { content, _id: commentId } = comment
+  const saveUpdatedComment = async e => {
+    if (e.key === 'Enter') {
+      const commentUpdates = {
+        content: editedComment,
+      }
+      await updateComment(commentId, commentUpdates)
+      toggleFetchComments(state => !state)
+      toggleEditMode(false)
+    }
+  }
+
+  const handleEditCommentChange = e => setEditedComment(e.target.value)
+
   return (
     <div className='comment-content'>
       {editMode ? (
@@ -12,7 +29,7 @@ export const CommentContent = ({
           className='edit-comment'
           type='text'
           value={editedComment}
-          onChange={e => setEditedComment(e.target.value)}
+          onChange={handleEditCommentChange}
           onKeyUp={saveUpdatedComment}
         />
       ) : (
