+++ src/screens/Project/TaskManagement/TicketDialog/Comments/CommentFooter.tsx
@@ -1,23 +1,54 @@
 import { BiLike } from 'react-icons/bi'
+import { deleteComment, updateComment } from 'utils/api/tickets'
+import { useAppSelector } from 'utils/redux/hooks'
+import { selectUserId } from 'utils/redux/slices/userSlice'
 
 export const CommentFooter = ({
-  author,
-  currentUser,
-  isReply,
-  renderReplyInputBar,
-  allowEditComment,
-  deleteThisComment,
-  likeComment,
-  likes,
+  comment,
+  toggleFetchComments,
+  fetchComments,
+  toggleEditMode,
+  toggleRenderReplyInput,
 }) => {
+  const userId = useAppSelector(selectUserId)
+  const { author, isReply, likes, _id: commentId } = comment
+  const allowEditComment = () => {
+    toggleEditMode(state => !state)
+  }
+
+  const deleteThisComment = async () => {
+    await deleteComment(commentId)
+    toggleFetchComments(!fetchComments)
+  }
+
+  const likeComment = async () => {
+    let updatedLikes
+    if (likes.includes(userId)) {
+      const idx = likes.indexOf(userId)
+      updatedLikes = {
+        likes: likes.splice(idx, 1),
+      }
+    } else {
+      updatedLikes = {
+        likes: [...likes, userId],
+      }
+    }
+    await updateComment(commentId, updatedLikes)
+    toggleFetchComments(!fetchComments)
+  }
+
+  const renderReplyInputBar = () => {
+    toggleRenderReplyInput(true)
+  }
+
   return (
     <div className='comment-actions'>
-      {author.userId !== currentUser._id && !isReply && (
+      {author.userId !== userId && !isReply && (
         <div onClick={renderReplyInputBar}>
           <p>Reply</p>
         </div>
       )}
-      {author.userId === currentUser._id && (
+      {author.userId === userId && (
         <>
           <div onClick={allowEditComment}>
             <p>Edit</p>
